Split fullMatrix class to own file
